==> notebooks/02_download_electoral_data.ipynb <==
# Setup

import pandas as pd

# Read data

def read_data():
    url = 'https://en.wikipedia.org/wiki/Electoral_College_(United_States)'
    
    return (
        pd.read_html(url)
        [4]
        .loc[4:54, [1, 35]]
        .rename(columns={1: 'state', 35: 'electors'})
    )

# Clean data

def clean_state(df):
    return df.state.replace('D.C.', 'District of Columbia')

def clean_electors(df):
    return df.electors.astype(int)

def clean(df):
    return df.assign(
        state=clean_state,
        electors=clean_electors
    )

# Add columns

def assign_pct_of_electors(df):
    return df.electors.div(df.electors.sum()).mul(100).round(2)

def assign(df):
    return df.assign(
        pct_of_electors=assign_pct_of_electors
    )

# Remove columns

def remove(df):
    return df.drop('electors', axis=1)

# Putting it together

df = read_data().pipe(clean).pipe(assign).pipe(remove).set_index('state').sort_index()

# To CSV

df.to_csv('../data/electors.csv')
