==> notebooks/03_download_election.ipynb <==
# Setup

import pandas as pd

# Read data

def read_data():
    url = 'https://en.wikipedia.org/wiki/United_States_elections,_2016'
    return (
        pd.read_html(url, header=1)
        [1]
        .drop([50, 51, 58, 59])
        .loc[:, ['Subdivision', 'Pres.']]
        .rename(columns={'Subdivision': 'state', 'Pres.': 'president'})
    )

# Clean data

def clean_state(df):
    return df.state.replace('Washington, D.C.', 'District of Columbia')

def clean(df):
    return df.assign(
        state=clean_state
    )

# Putting it together

df = read_data().pipe(clean).set_index('state').sort_index()

# To CSV

df.to_csv('../data/president.csv')
