==> notebooks/04_combine_data.ipynb <==
# Setup

import pandas as pd

# Load data

def load(filename):
    return pd.read_csv(filename, index_col=['state']).sort_index()

def load_data():
    pop = load('../data/pop.csv')
    electors = load('../data/electors.csv')
    president = load('../data/president.csv')
    
    return pop.join(electors, how='inner').join(president, how='inner')

df = load_data()

# To CSV

df.to_csv('../data/final.csv')
